+++ examples/view-locals/index.js
@@ -88,7 +88,6 @@ function count2(req, res, next) {
 function users2(req, res, next) {
   User.all(function(err, users){
     if (err) return next(err);
-    // this would not be ideal for *this*
     res.locals.users = users.filter(ferrets);
     next();
   })
@@ -103,31 +102,50 @@ app.get('/middleware-locals', count2, users2, function(req, res, next){
 });
 
 
-
-
-// let's assume we wanted to load the users
-// and count for every res.render() call, we
-// could use app.locals.use() for this. These
-// are callbacks which run in parallel ONLY
-// when res.render() is invoked. If no views
-// are rendered, there is no overhead.
-
-// This may be ideal if you want to load auxiliary
-// user information, but only for templates. Note
-// that (req, res) are available to you, so you may
-// access req.session.user etc.
-
-// Keep in mind these execute in *parallel*, so these
-// callbacks should not depend on each other, this
-// also makes them slightly more efficient than
-// using middleware which execute sequentially
-
-app.locals.use(count2);
-app.locals.use(users2);
-
 app.get('/locals', function(req, res){
   res.render('user', { title: 'Users' });
 });
+
+// keep in mind that middleware may be placed anywhere
+// and in various combinations, so if you have locals
+// that you wish to make available to all subsequent
+// middleware/routes you can do something like this:
+
+/*
+
+app.use(function(req, res, next){
+  res.locals.user = req.user;
+  res.locals.sess = req.session;
+  next();
+});
+
+*/
+
+// or suppose you have some /admin
+// "global" local variables:
+
+/*
+
+app.use('/api', function(req, res, next){
+  res.locals.user = req.user;
+  res.locals.sess = req.session;
+  next();
+});
+
+*/
+
+// the following is effectively the same,
+// but uses a route instead:
+
+/*
+
+app.all('/api/*', function(req, res, next){
+  res.locals.user = req.user;
+  res.locals.sess = req.session;
+  next();
+});
+
+*/
 
 app.listen(3000);
 console.log('Application listening on port 3000');