+++ test/res.json.js
@@ -15,13 +15,30 @@ describe('res', function(){
 
         request(app)
         .get('/?callback=something')
-        .end(function(res){
+        .end(function(err, res){
           res.headers.should.have.property('content-type', 'text/javascript; charset=utf-8');
-          res.body.should.equal('something({"count":1});');
+          res.text.should.equal('something({"count":1});');
           done();
         })
       })
 
+      it('should allow renaming callback', function(done){
+        var app = express();
+        
+        app.set('jsonp callback name', 'clb');
+        app.use(function(req, res){
+          res.json({ count: 1 });
+        });
+
+        request(app)
+        .get('/?clb=something')
+        .end(function(err, res){
+          res.headers.should.have.property('content-type', 'text/javascript; charset=utf-8');
+          res.text.should.equal('something({"count":1});');
+          done();
+        })
+      })      
+      
       it('should allow []', function(done){
         var app = express();
 
@@ -31,9 +48,9 @@ describe('res', function(){
 
         request(app)
         .get('/?callback=callbacks[123]')
-        .end(function(res){
+        .end(function(err, res){
           res.headers.should.have.property('content-type', 'text/javascript; charset=utf-8');
-          res.body.should.equal('callbacks[123]({"count":1});');
+          res.text.should.equal('callbacks[123]({"count":1});');
           done();
         })
       })
@@ -49,9 +66,9 @@ describe('res', function(){
 
         request(app)
         .get('/')
-        .end(function(res){
+        .end(function(err, res){
           res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
-          res.body.should.equal('null');
+          res.text.should.equal('null');
           done();
         })
       })
@@ -67,9 +84,9 @@ describe('res', function(){
 
         request(app)
         .get('/')
-        .end(function(res){
+        .end(function(err, res){
           res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
-          res.body.should.equal('["foo","bar","baz"]');
+          res.text.should.equal('["foo","bar","baz"]');
           done();
         })
       })
@@ -85,9 +102,9 @@ describe('res', function(){
 
         request(app)
         .get('/')
-        .end(function(res){
+        .end(function(err, res){
           res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
-          res.body.should.equal('{"name":"tobi"}');
+          res.text.should.equal('{"name":"tobi"}');
           done();
         })
       })
@@ -109,8 +126,8 @@ describe('res', function(){
 
         request(app)
         .get('/')
-        .end(function(res){
-          res.body.should.equal('{"name":"tobi"}');
+        .end(function(err, res){
+          res.text.should.equal('{"name":"tobi"}');
           done();
         });
       })
@@ -140,8 +157,8 @@ describe('res', function(){
 
         request(app)
         .get('/')
-        .end(function(res){
-          res.body.should.equal('{\n  "name": "tobi",\n  "age": 2\n}');
+        .end(function(err, res){
+          res.text.should.equal('{\n  "name": "tobi",\n  "age": 2\n}');
           done();
         });
       })
@@ -158,10 +175,29 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.statusCode.should.equal(201);
         res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
-        res.body.should.equal('{"id":1}');
+        res.text.should.equal('{"id":1}');
+        done();
+      })
+    })
+  })
+
+  describe('.json(object, status)', function(){
+    it('should respond with json and set the .statusCode for backwards compat', function(done){
+      var app = express();
+
+      app.use(function(req, res){
+        res.json({ id: 1 }, 201);
+      });
+
+      request(app)
+      .get('/')
+      .end(function(err, res){
+        res.statusCode.should.equal(201);
+        res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
+        res.text.should.equal('{"id":1}');
         done();
       })
     })