+++ test/res.redirect.js
@@ -4,26 +4,6 @@ var express = require('../')
 
 describe('res', function(){
   describe('.redirect(url)', function(){
-    it('should respect X-Forwarded-Proto when "trust proxy" is enabled', function(done){
-      var app = express();
-
-      app.enable('trust proxy');
-
-      app.use(function(req, res){
-        res.redirect('/login');
-      });
-
-      request(app)
-      .get('/')
-      .set('Host', 'example.com')
-      .set('X-Forwarded-Proto', 'https')
-      .end(function(res){
-        res.statusCode.should.equal(302);
-        res.headers.should.have.property('location', 'https://example.com/login');
-        done();
-      })
-    })
-
     it('should default to a 302 redirect', function(done){
       var app = express();
 
@@ -33,15 +13,34 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.statusCode.should.equal(302);
         res.headers.should.have.property('location', 'http://google.com');
         done();
       })
     })
+
+    describe('with leading //', function(){
+      it('should pass through scheme-relative urls', function(done){
+        var app = express();
+
+        app.use(function(req, res){
+          res.redirect('//cuteoverload.com');
+        });
+
+        request(app)
+        .get('/')
+        .set('Host', 'example.com')
+        .end(function(err, res){
+          res.headers.should.have.property('location', '//cuteoverload.com');
+          done();
+        })
+      })
+    })
+
     
     describe('with leading /', function(){
-      it('should construct host-relative urls', function(done){
+      it('should construct scheme-relative urls', function(done){
         var app = express();
 
         app.use(function(req, res){
@@ -51,8 +50,8 @@ describe('res', function(){
         request(app)
         .get('/')
         .set('Host', 'example.com')
-        .end(function(res){
-          res.headers.should.have.property('location', 'http://example.com/login');
+        .end(function(err, res){
+          res.headers.should.have.property('location', '//example.com/login');
           done();
         })
       })
@@ -69,8 +68,8 @@ describe('res', function(){
         request(app)
         .get('/post/1')
         .set('Host', 'example.com')
-        .end(function(res){
-          res.headers.should.have.property('location', 'http://example.com/post/1/./edit');
+        .end(function(err, res){
+          res.headers.should.have.property('location', '//example.com/post/1/./edit');
           done();
         })
       })
@@ -87,8 +86,8 @@ describe('res', function(){
         request(app)
         .get('/post/1')
         .set('Host', 'example.com')
-        .end(function(res){
-          res.headers.should.have.property('location', 'http://example.com/post/1/../new');
+        .end(function(err, res){
+          res.headers.should.have.property('location', '//example.com/post/1/../new');
           done();
         })
       })
@@ -105,8 +104,8 @@ describe('res', function(){
         request(app)
         .get('/')
         .set('Host', 'example.com')
-        .end(function(res){
-          res.headers.should.have.property('location', 'http://example.com/login');
+        .end(function(err, res){
+          res.headers.should.have.property('location', '//example.com/login');
           done();
         })
       })
@@ -129,8 +128,8 @@ describe('res', function(){
           request(app)
           .get('/blog/admin')
           .set('Host', 'example.com')
-          .end(function(res){
-            res.headers.should.have.property('location', 'http://example.com/blog/admin/login');
+          .end(function(err, res){
+            res.headers.should.have.property('location', '//example.com/blog/admin/login');
             done();
           })
         })
@@ -150,8 +149,8 @@ describe('res', function(){
           request(app)
           .get('/blog')
           .set('Host', 'example.com')
-          .end(function(res){
-            res.headers.should.have.property('location', 'http://example.com/blog/admin/login');
+          .end(function(err, res){
+            res.headers.should.have.property('location', '//example.com/blog/admin/login');
             done();
           })
         })
@@ -171,8 +170,8 @@ describe('res', function(){
           request(app)
           .get('/blog')
           .set('Host', 'example.com')
-          .end(function(res){
-            res.headers.should.have.property('location', 'http://example.com/admin/login');
+          .end(function(err, res){
+            res.headers.should.have.property('location', '//example.com/admin/login');
             done();
           })
         })
@@ -190,7 +189,7 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.statusCode.should.equal(303);
         res.headers.should.have.property('location', 'http://google.com');
         done();
@@ -208,9 +207,9 @@ describe('res', function(){
 
       request(app)
       .head('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('location', 'http://google.com');
-        res.body.should.equal('');
+        res.text.should.equal('');
         done();
       })
     })
@@ -227,9 +226,9 @@ describe('res', function(){
       request(app)
       .get('/')
       .set('Accept', 'text/html')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('location', 'http://google.com');
-        res.body.should.equal('<p>Moved Temporarily. Redirecting to <a href="http://google.com">http://google.com</a></p>');
+        res.text.should.equal('<p>Moved Temporarily. Redirecting to <a href="http://google.com">http://google.com</a></p>');
         done();
       })
     })
@@ -246,10 +245,10 @@ describe('res', function(){
       request(app)
       .get('/')
       .set('Accept', 'text/plain, */*')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('location', 'http://google.com');
         res.headers.should.have.property('content-length', '51');
-        res.body.should.equal('Moved Temporarily. Redirecting to http://google.com');
+        res.text.should.equal('Moved Temporarily. Redirecting to http://google.com');
         done();
       })
     })
@@ -266,12 +265,12 @@ describe('res', function(){
       request(app)
       .get('/')
       .set('Accept', 'foo/bar')
-      .end(function(res){
+      .end(function(err, res){
         res.should.have.status(302);
         res.headers.should.have.property('location', 'http://google.com');
         res.headers.should.not.have.property('content-type');
         res.headers.should.have.property('content-length', '0');
-        res.body.should.equal('');
+        res.text.should.equal('');
         done();
       })
     })