+++ test/res.send.js
@@ -13,10 +13,7 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.body.should.equal('');
-        done();
-      })
+      .expect('', done);
     })
   })
   
@@ -30,10 +27,7 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.body.should.equal('');
-        done();
-      })
+      .expect('', done);
     })
   })
 
@@ -47,11 +41,8 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.body.should.equal('Created');
-        res.statusCode.should.equal(201);
-        done();
-      })
+      .expect('Created')
+      .expect(201, done);
     })
   })
   
@@ -65,14 +56,26 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.body.should.equal('Created :)');
-        res.statusCode.should.equal(201);
-        done();
-      })
+      .expect('Created :)')
+      .expect(201, done);
     })
   })
-  
+
+  describe('.send(body, code)', function(){
+    it('should be supported for backwards compat', function(done){
+      var app = express();
+
+      app.use(function(req, res){
+        res.send('Bad!', 400);
+      });
+
+      request(app)
+      .get('/')
+      .expect('Bad!')
+      .expect(400, done);
+    })
+  })
+
   describe('.send(String)', function(){
     it('should send as html', function(done){
       var app = express();
@@ -83,9 +86,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'text/html; charset=utf-8');
-        res.body.should.equal('<p>hey</p>');
+        res.text.should.equal('<p>hey</p>');
         res.statusCode.should.equal(200);
         done();
       })
@@ -101,10 +104,8 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.headers.should.have.property('etag', '-1498647312');
-        done();
-      })
+      .expect('ETag', '-1498647312')
+      .end(done);
     })
     
     it('should not override Content-Type', function(done){
@@ -116,12 +117,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.headers.should.have.property('content-type', 'text/plain');
-        res.body.should.equal('hey');
-        res.statusCode.should.equal(200);
-        done();
-      })
+      .expect('Content-Type', 'text/plain')
+      .expect('hey')
+      .expect(200, done);
     })
   })
   
@@ -135,9 +133,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'text/html; charset=utf-8');
-        res.body.should.equal('<p>hey</p>');
+        res.text.should.equal('<p>hey</p>');
         res.statusCode.should.equal(200);
         done();
       })
@@ -152,9 +150,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'text/plain');
-        res.body.should.equal('hey');
+        res.text.should.equal('hey');
         res.statusCode.should.equal(200);
         done();
       })
@@ -171,9 +169,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'application/octet-stream');
-        res.body.should.equal('hello');
+        res.text.should.equal('hello');
         res.statusCode.should.equal(200);
         done();
       })
@@ -189,10 +187,8 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
-        res.headers.should.have.property('etag', '-1498647312');
-        done();
-      })
+      .expect('ETag', '-1498647312')
+      .end(done);
     })
 
     it('should not override Content-Type', function(done){
@@ -204,9 +200,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'text/plain');
-        res.body.should.equal('hey');
+        res.text.should.equal('hey');
         res.statusCode.should.equal(200);
         done();
       })
@@ -223,9 +219,9 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
-        res.body.should.equal('{"name":"tobi"}');
+        res.text.should.equal('{"name":"tobi"}');
         done();
       })
     })
@@ -241,10 +237,7 @@ describe('res', function(){
 
       request(app)
       .head('/')
-      .end(function(res){
-        res.body.should.equal('');
-        done();
-      })
+      .expect('', done);
     })
   })
 
@@ -258,10 +251,10 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.not.have.property('content-type');
         res.headers.should.not.have.property('content-length');
-        res.body.should.equal('');
+        res.text.should.equal('');
         done();
       })
     })
@@ -277,10 +270,10 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .end(function(res){
+      .end(function(err, res){
         res.headers.should.not.have.property('content-type');
         res.headers.should.not.have.property('content-length');
-        res.body.should.equal('');
+        res.text.should.equal('');
         done();
       })
     })
@@ -326,10 +319,7 @@ describe('res', function(){
     request(app)
     .get('/')
     .set('If-None-Match', 'asdf')
-    .end(function(res){
-      res.should.have.status(500);
-      res.body.should.equal('hey');
-      done();
-    });
+    .expect('hey')
+    .expect(500, done);
   })
 })